Implement functionality to calculate average time spent on word (only in alphabet feature)

+++ app/src/main/java/yaroslavgorbach/koropapps/vocabulary/data/statistics/local/model/SttisticsTimeEntity.kt
@@ -2,7 +2,6 @@ package yaroslavgorbach.koropapps.vocabulary.data.statistics.local.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import yaroslavgorbach.koropapps.vocabulary.data.exercises.local.model.ExerciseName
 import java.util.*
 
 @Entity
@@ -10,6 +9,6 @@ data class StatisticsTimeEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
     val exerciseNameRes: Int,
-    val value: Int,
+    val value: Float,
     val date: Date
 )
